Move keymap operations to ShortcutsList
